+++ mackup/constants.py
@@ -33,4 +33,5 @@ CUSTOM_APPS_DIR = '.mackup'
 # Supported engines
 ENGINE_DROPBOX = 'dropbox'
 ENGINE_GDRIVE = 'google_drive'
+ENGINE_COPY = 'copy'
 ENGINE_FS = 'file_system'